// src/store/types/proposal.actionTypes.ts
export enum PROPOSAL_ACTION_TYPES {
  // Get Proposals by RFP
  GET_PROPOSALS_BY_RFP_REQUEST = 'GET_PROPOSALS_BY_RFP_REQUEST',
  GET_PROPOSALS_BY_RFP_SUCCESS = 'GET_PROPOSALS_BY_RFP_SUCCESS',
  GET_PROPOSALS_BY_RFP_FAILURE = 'GET_PROPOSALS_BY_RFP_FAILURE',

  // Get Proposal by ID
  GET_PROPOSAL_BY_ID_REQUEST = 'GET_PROPOSAL_BY_ID_REQUEST',
  GET_PROPOSAL_BY_ID_SUCCESS = 'GET_PROPOSAL_BY_ID_SUCCESS',
  GET_PROPOSAL_BY_ID_FAILURE = 'GET_PROPOSAL_BY_ID_FAILURE',

  // Create Proposal
  CREATE_PROPOSAL_REQUEST = 'CREATE_PROPOSAL_REQUEST',
  CREATE_PROPOSAL_SUCCESS = 'CREATE_PROPOSAL_SUCCESS',
  CREATE_PROPOSAL_FAILURE = 'CREATE_PROPOSAL_FAILURE',

  // Update Proposal
  UPDATE_PROPOSAL_REQUEST = 'UPDATE_PROPOSAL_REQUEST',
  UPDATE_PROPOSAL_SUCCESS = 'UPDATE_PROPOSAL_SUCCESS',
  UPDATE_PROPOSAL_FAILURE = 'UPDATE_PROPOSAL_FAILURE',

  // Delete Proposal
  DELETE_PROPOSAL_REQUEST = 'DELETE_PROPOSAL_REQUEST',
  DELETE_PROPOSAL_SUCCESS = 'DELETE_PROPOSAL_SUCCESS',
  DELETE_PROPOSAL_FAILURE = 'DELETE_PROPOSAL_FAILURE',

  // Parse Email to Proposal
  PARSE_EMAIL_TO_PROPOSAL_REQUEST = 'PARSE_EMAIL_TO_PROPOSAL_REQUEST',
  PARSE_EMAIL_TO_PROPOSAL_SUCCESS = 'PARSE_EMAIL_TO_PROPOSAL_SUCCESS',
  PARSE_EMAIL_TO_PROPOSAL_FAILURE = 'PARSE_EMAIL_TO_PROPOSAL_FAILURE',

  // Simulate Vendor Response
  SIMULATE_VENDOR_RESPONSE_REQUEST = 'SIMULATE_VENDOR_RESPONSE_REQUEST',
  SIMULATE_VENDOR_RESPONSE_SUCCESS = 'SIMULATE_VENDOR_RESPONSE_SUCCESS',
  SIMULATE_VENDOR_RESPONSE_FAILURE = 'SIMULATE_VENDOR_RESPONSE_FAILURE',

  // Clear Errors
  CLEAR_PROPOSAL_ERRORS = 'CLEAR_PROPOSAL_ERRORS',
}